Compile dtksh on FreeBSD
dtksh can be now compiled on FreeBSD. Work in progress.

Needs a real Korn shell to bootstrap as $(KSHELL).
KSHELL is set by default to /usr/local/bin/ksh93
(generic POSIX shell may not work)

Tested on:

  FreeBSD 9.0-BETA1 #0 r224912M amd64

Known issues:

xvmstat:
* sleep does not work well (SIGSTOP is delivered)
xpong:
* xpong: line 220:  ball1x = max_x * 2.2 / 3 : arithmetic syntax error
* dtksh is rebuilt uncondtionally every time make is invoked

+++ cde/programs/dtksh/ksh93/ship/shipproto.c
@@ -104,6 +104,7 @@
 
 # include <sys/types.h>
 # include <sys/stat.h>
+# include <string.h>
 
 # undef	ctime
 
@@ -1331,7 +1332,7 @@ struct ppsymkey
 
 
 extern __MANGLE__ int		access __PROTO__((const char*, int));
-extern __MANGLE__ __V_*		calloc __PROTO__((int, int));
+extern __MANGLE__ __V_*		calloc __PROTO__((size_t, size_t));
 extern __MANGLE__ int		close __PROTO__((int));
 extern __MANGLE__ int		creat __PROTO__((const char*, int));
 extern __MANGLE__ char*		ctime __PROTO__((time_t*));
